#  Make tuple t[2:end] inferrable through constant prop