Spurred by #31132, I was curious if our constant propagation was robust enough to handle a peephole optimization here -- and it is! This removes the need to export front/tail by simply allowing folks to get inferrable results when they spell these operations as `t[2:end]` and `t[1:end-1]`.  I additionally allowed for `t[3:end]` and `t[1:end-2]` (as well as the trivial `t[1:end]`) as I have used `tail(tail(t))` in the past.

cc @timholy — this is the spiritual successor to https://github.com/JuliaLang/julia/pull/20370. Two years later and we have our holy grail!