@mbauman 2019-02-22
Oh thanks for the pointer @tkf — I had somehow missed that issue/PR.  This implementation should be much less stressing on the compiler.  It's less general, too, but it should hit the most common uses of a constant value.