@timholy 2020-01-22
I'd like to merge this, but just want to point out that it doesn't fix *all* tuple indexing, it just special-cases the 4 situations @mbauman mentions, `t[2:end]`, `t[3:end]`, `t[1:end-1]`, and `t[1:end-2]`. So the counter-argument is that by optimizing these cases we make it *more* surprising when this doesn't work in general. I am personally fine with that, as I think these cases make up the majority of uses.

But I'll leave this open for another couple of days in case people disagree.