@perrutquist 2020-02-20
I recently wrote a proposal on discourse about [branching on whether a value is known to the compiler](https://discourse.julialang.org/t/branch-on-whether-a-value-is-known-to-the-compiler/34850). 
I don't know enough about the compiler internals to even say if it's possible to implement that, but if the feature existed then it'd be very easy to handle tuple indexing in the general case.