@mbauman 2020-04-29
I've rebased this and ran a few quick benchmarks just to make sure this is still needed and doing what we want.  It is.

Before:

```Julia
julia> f(t) = t[2:end]
f (generic function with 1 method)

julia> @btime f($(1,2,3,))
  149.206 ns (2 allocations: 128 bytes)
(2, 3)

julia> @btime f($(1,2,3,4,5,6,7,8))
  261.398 ns (2 allocations: 208 bytes)
(2, 3, 4, 5, 6, 7, 8)

julia> @btime f($(1,"two",3.0,4//1,0x5,0x0006))
  286.491 ns (9 allocations: 544 bytes)
("two", 3.0, 4//1, 0x05, 0x0006)

julia> f(t) = t[3:5]
f (generic function with 1 method)

julia> @btime f($(1,2,3,4,5,6,7,8))
  171.353 ns (2 allocations: 144 bytes)
(3, 4, 5)

julia> @btime f($(1,"two",3.0,4//1,0x5,0x0006))
  222.964 ns (7 allocations: 400 bytes)
(3.0, 4//1, 0x05)
```

After:

```julia
julia> f(t) = t[2:end]
f (generic function with 1 method)

julia> @btime f($(1,2,3,))
  0.035 ns (0 allocations: 0 bytes)
(2, 3)

julia> @btime f($(1,2,3,4,5,6,7,8))
  0.035 ns (0 allocations: 0 bytes)
(2, 3, 4, 5, 6, 7, 8)

julia> @btime f($(1,"two",3.0,4//1,0x5,0x0006))
  2.366 ns (0 allocations: 0 bytes)
("two", 3.0, 4//1, 0x05, 0x0006)

julia> f(t) = t[3:5] # still not special-cased
f (generic function with 1 method)

julia> @btime f($(1,2,3,4,5,6,7,8))
  172.339 ns (2 allocations: 144 bytes)
(3, 4, 5)

julia> @btime f($(1,"two",3.0,4//1,0x5,0x0006))
  225.719 ns (7 allocations: 400 bytes)
(3.0, 4//1, 0x05)
```

`@code_warntype` also shows what you'd expect — with all the downstream impacts that that has.